fix(schema): Execution token 累计列迁 BigInt，修 int32 溢出导致链路追踪整页拉不出
Execution 的六个 token 计数列是整个 session 累计求和，用 Prisma Int（32 位，上限
2,147,483,647）存不下长命 session。溢出后 Prisma 抛 P2023，此后任何 select 到这些
列的查询整条失败 —— 不止坏溢出的那一行，那一批 trace 全部拉不出。

实测确认了超限行是怎么进到库里的：create/upsert 确实抛 P2023，但抛错发生在 INSERT
提交之后、Prisma 回读返回值那一步，行已经落库、值就是超限值。

改动：
- schema：tokens/inputTokens/outputTokens/cacheReadInputTokens/
  cacheCreationInputTokens/reasoningTokens 六列 Int? → BigInt?。
  maxSingleCallTokens 是单次取 max、不累计，保持 Int。
- 读边界收口：新增 src/lib/storage/prisma-client.ts，用 Prisma client extension 把
  这六列转回 number。Prisma 读 BigInt 列一律返回 bigint（连 42 也是），而 bigint 既
  不能 JSON.stringify 也不能与 number 混算；全站 200+ 处按 number 消费这些字段，
  收口后下游一行未改。Number 对 2^53-1 内整数精确，是 32 位上限的 400 万倍，无损。
  null 保持 null —— 转成 0 会把「没采到」伪装成「用了 0 token」。
- 写入侧不变：Prisma 的 BigInt 列接受 number 入参，现有写入点零改动。
- 客户端创建收敛到共用 createPrismaClient()（原先 prisma.ts 与 db-interface.ts 各
  new 一个，容易分叉）。
- 启动迁移：新增 scripts/db_push.sh，由三条启动路径共用。db push 不带
  --accept-data-loss 时一遇到列类型变更就退出码 1，直接合入会让重启失败、服务起不来。
  包装器只在告警全部是 Int→BigInt 时才放行，其余破坏性变更照旧拦下；并统一非交互
  运行，避免 TTY 下 "reset database?" 回车默认 Yes 清库。
- openGauss：init_opengauss.py 同一批列 INTEGER → BIGINT。该脚本只建新表/补缺失列、
  不改已存在列的类型，老实例需手动 ALTER（已在注释里写明）。

验证：
- 新增 test/execution-token-bigint.test.ts 6 条：超限值往返、读回是 number、null 不
  变 0、一行溢出不再拖垮整条 findMany、未迁移旧库上正常值照常读写、未迁移旧库上修复
  不生效（把「部署必须迁移」这个前提钉死）。
- 真实库副本实跑迁移：行数 1087→1087、tokens 合计 139,733,994 不变、逐行指纹一致、
  6 个索引都在。
- tsc 0 错；全量 685 tests / 675 pass / 9 fail，与基线 679/669/9 逐条对齐（那 9 条是
  该工作树缺 DB 的 experiment 系列，报 "Unable to open the database file"，与本改动
  无关）。

Fixes #249

+++ scripts/init_opengauss.py
@@ -61,7 +61,12 @@ TABLE_DEFINITIONS = {
             ("taskId", "TEXT"),
             ("query", "TEXT"),
             ("framework", "TEXT"),
-            ("tokens", "INTEGER"),
+            # 下面几列（tokens/inputTokens/outputTokens/cacheRead*/cacheCreation*）是整个
+            # session 累计求和，长命 session 会突破 int4 的 2,147,483,647，故用 BIGINT，
+            # 与 prisma/schema.prisma 的 BigInt 保持一致。
+            # 注意：本脚本只建新表 / 给缺失列做 ADD COLUMN，不会改已存在列的类型。已经建过库的
+            # openGauss 实例需要手动 ALTER TABLE "Execution" ALTER COLUMN "<列>" TYPE BIGINT。
+            ("tokens", "BIGINT"),
             ("cost", "FLOAT"),
             ("latency", "FLOAT"),
             ("timestamp", "TIMESTAMP DEFAULT CURRENT_TIMESTAMP"),
@@ -81,11 +86,12 @@ TABLE_DEFINITIONS = {
             ("user", "TEXT"),
             ("toolCallCount", "INTEGER"),
             ("llmCallCount", "INTEGER"),
-            ("inputTokens", "INTEGER"),
-            ("outputTokens", "INTEGER"),
+            ("inputTokens", "BIGINT"),
+            ("outputTokens", "BIGINT"),
             ("toolCallErrorCount", "INTEGER"),
-            ("cacheReadInputTokens", "INTEGER"),
-            ("cacheCreationInputTokens", "INTEGER"),
+            ("cacheReadInputTokens", "BIGINT"),
+            ("cacheCreationInputTokens", "BIGINT"),
+            # 单次调用取 max、不累计，到不了 int4 上限，保持 INTEGER。
             ("maxSingleCallTokens", "INTEGER"),
         ],
         "unique_constraints": []
